# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nutri/patients/{patientId}/plans": {
      "get": {
        "description": "Sem o grafo, com o TAMANHO dele (dayTypeCount, mealCount) e `semanaCompleta` — plano cuja semana não está programada não serve ao app do paciente. Ativo primeiro, depois o mais recente.",
        "operationId": "PlanoController_listarPlanos",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "planos do paciente"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "paciente não encontrado"
          }
        },
        "summary": "Listar os planos do paciente",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "post": {
        "description": "O plano nasce VAZIO: nem tipo-de-dia, nem refeição, nem programação de semana. O PRIMEIRO plano do paciente nasce ativo (não existe estado \"nenhum plano ativo\"); do segundo em diante nasce inativo, porque trocar o plano ativo é o ato que o ciclo observa (POST /nutri/patients/:id/active-plan) e não pode ser efeito colateral de um cadastro.",
        "operationId": "PlanoController_criarPlano",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Plano de julho"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "o plano criado, com o grafo vazio"
          },
          "400": {
            "description": "name ausente, vazio ou longo demais"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "paciente não encontrado"
          }
        },
        "summary": "Criar um plano",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/plans/{planId}": {
      "get": {
        "description": "Tipos-de-dia, programação da semana, refeições (ordenadas por posição), opções (a padrão primeiro) e itens com nome do alimento e do grupo já resolvidos. É a leitura que a tela do editor faz — uma, não uma por nó.",
        "operationId": "PlanoController_detalhePlano",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "o grafo do plano"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "plano não encontrado"
          }
        },
        "summary": "O plano inteiro em uma requisição",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "patch": {
        "description": "Só o nome. `isActive` NÃO é editável aqui: ativar plano é o ato observado pelo ciclo e continua em POST /nutri/patients/:patientId/active-plan (007).",
        "operationId": "PlanoController_atualizarPlano",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Plano de agosto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "o plano depois da alteração"
          },
          "400": {
            "description": "name vazio ou longo demais"
          },
          "404": {
            "description": "plano não encontrado"
          }
        },
        "summary": "Renomear o plano",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "delete": {
        "description": "Cascata para baixo (tipos-de-dia, refeições, opções, itens, semana). RECUSA com 409 em três casos: há registro de refeição no plano; há vigência dele em algum ciclo; é o plano ativo de um paciente com ciclo aberto.",
        "operationId": "PlanoController_excluirPlano",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "plano e grafo excluídos"
          },
          "404": {
            "description": "plano não encontrado"
          },
          "409": {
            "description": "registro, vigência de ciclo, ou plano ativo com ciclo aberto"
          }
        },
        "summary": "Excluir o plano e o grafo abaixo",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/plans/{planId}/schedule": {
      "put": {
        "description": "A semana é UM objeto, não 7 linhas independentes: uma semana com 6 dias programados é um estado inválido. O PUT substitui a programação inteira. `days`: lista de {weekday 0=domingo…6=sábado, dayTypeId}, exatamente os 7 dias, sem repetir. Tipo-de-dia de outro plano → 422.",
        "operationId": "PlanoController_definirSemana",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "days": [
                    {
                      "weekday": 0,
                      "dayTypeId": "…"
                    },
                    {
                      "weekday": 1,
                      "dayTypeId": "…"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "o plano com a semana programada"
          },
          "400": {
            "description": "days não é lista, não cobre os 7 dias, ou repete um dia"
          },
          "404": {
            "description": "plano não encontrado"
          },
          "422": {
            "description": "tipo-de-dia não pertence a este plano"
          }
        },
        "summary": "Programar a semana (os 7 dias de uma vez)",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/plans/{planId}/day-types": {
      "post": {
        "description": "Nasce VAZIO (sem refeição). O plano é um CONJUNTO de tipos-de-dia (treino / descanso), não um cardápio fixo.",
        "operationId": "PlanoController_criarTipoDia",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Treino"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "o tipo-de-dia criado, vazio"
          },
          "400": {
            "description": "name ausente ou vazio"
          },
          "404": {
            "description": "plano não encontrado"
          }
        },
        "summary": "Criar um tipo-de-dia",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/day-types/{dayTypeId}": {
      "patch": {
        "operationId": "PlanoController_atualizarTipoDia",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dayTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Descanso"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "o tipo-de-dia depois da alteração"
          },
          "404": {
            "description": "tipo-de-dia não encontrado"
          }
        },
        "summary": "Renomear o tipo-de-dia",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "delete": {
        "description": "Cascata em refeições, opções e itens. RECUSA com 409 se a programação da semana o referencia (a semana ficaria com um dia sem tipo) ou se há registro em alguma refeição sua.",
        "operationId": "PlanoController_excluirTipoDia",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dayTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "tipo-de-dia excluído"
          },
          "404": {
            "description": "tipo-de-dia não encontrado"
          },
          "409": {
            "description": "está na semana, ou tem registro"
          }
        },
        "summary": "Excluir o tipo-de-dia e o que pende abaixo",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bamboo API",
    "description": "API da alça do paciente — ver \"o agora\" (US1) e substituir um alimento dentro do grupo, com quantidade recalculada e medida caseira (US2). v0: auth stub (paciente fixo por env).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}